// src/components/ChannelCard.tsx
import React from 'react';
import { Box, CardContent, CardMedia, Typography } from '@mui/material';
import CheckCircleIcon from '@mui/icons-material/CheckCircle';
import { Link } from 'react-router-dom';
import { demoProfilePicture } from "../utilis/contants"
import { Item } from '../Types/types';

interface IProps{
  item:Item | undefined 
  marginTop:number | string

}

const ChannelCard = ({ item,marginTop}:IProps) => (
  <Box
    sx={{
      boxShadow: 'none',
      borderRadius: '20px',
      display: 'flex',
      justifyContent: 'center',
      alignItems: 'center',
      width: { xs: '356px', md: '320px' },
      height: '326px',
      margin: 'auto',
      marginTop,
    }}
  >
    <Link to={`/channel/${item?.id?.channelId}`}>
      <CardContent sx={{ display: 'flex', flexDirection: 'column', justifyContent: 'center', textAlign: 'center', color: '#fff' }}>
        <img
          src={item?.snippet?.thumbnails?.high?.url || demoProfilePicture}
          alt={item?.snippet?.title}
          style={{ borderRadius: '50%', height: '180px', width: '180px', marginBottom: 2, border: '1px solid #e3e3e3' }}
        />
        <Typography variant="h6">
          {item?.snippet?.title}{' '}
          <CheckCircleIcon sx={{ fontSize: '14px', color: 'gray', ml: '5px' }} />
        </Typography>
        {item?.statistics?.subscriberCount && (
          <Typography sx={{ fontSize: '15px', fontWeight: 500, color: 'gray' }}>
            {parseInt(item?.statistics?.subscriberCount).toLocaleString('en-US')} Subscribers
          </Typography>
        )}
      </CardContent>
    </Link>
  </Box>
);

export default ChannelCard;